+++ tests/doc/property/get_kind.cpp
@@ -9,9 +9,9 @@ using namespace chemfiles;
 
 TEST_CASE() {
     // [example]
-    assert(Property(42).get_kind() == Property::DOUBLE);
-    assert(Property(false).get_kind() == Property::BOOL);
-    assert(Property("string").get_kind() == Property::STRING);
-    assert(Property(Vector3D(1, 2, 3)).get_kind() == Property::VECTOR3D);
+    assert(Property(42).kind() == Property::DOUBLE);
+    assert(Property(false).kind() == Property::BOOL);
+    assert(Property("string").kind() == Property::STRING);
+    assert(Property(Vector3D(1, 2, 3)).kind() == Property::VECTOR3D);
     // [example]
 }